Progress in learning Unity 2D
Delete the temporary assets, create a folder to contain temporary assets and
add that folder to the git ignore.
Create a project for Tutsplus Arkanoid tuts series.
Create a project for Platformer Tutsplus tutorial, another one for the
sample project provided with the assets.
Create a folder to start doing a video series of unity 2D infinite runner.